+++ packages/core/src/generators/helpers/rsc.ts
@@ -1,7 +1,7 @@
 import { isMitosisNode } from '@/helpers/is-mitosis-node';
 import type { MitosisComponent } from '@/types/mitosis-component';
 import type { MitosisNode } from '@/types/mitosis-node';
-import traverse from 'traverse';
+import traverse from 'neotraverse/legacy';
 
 const checkIsNodeAMitosisComponent = (node: MitosisNode) =>
   node.name[0] === node.name[0].toUpperCase();