+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @builder.io/mitosis-cli
 
+## 0.3.14
+
+### Patch Changes
+
+- Updated dependencies [9a1d59b]
+  - @builder.io/mitosis@0.3.14
+
+## 0.3.13
+
+### Patch Changes
+
+- Updated dependencies [f86e2ec]
+  - @builder.io/mitosis@0.3.13
+
+## 0.3.12
+
+### Patch Changes
+
+- Updated dependencies [3a04558]
+  - @builder.io/mitosis@0.3.12
+
+## 0.3.11
+
+### Patch Changes
+
+- Updated dependencies [59a92da]
+  - @builder.io/mitosis@0.3.11
+
+## 0.3.10
+
+### Patch Changes
+
+- Updated dependencies [8548feb]
+- Updated dependencies [f83b8f4]
+  - @builder.io/mitosis@0.3.10
+
 ## 0.3.9
 
 ### Patch Changes